restructure: Add directory for logic related to AIR evaluation
