+++ triton-vm/src/air/tasm_air_constraints.rs
@@ -2,9 +2,10 @@
 //! Run `cargo run --bin constraint-evaluation-generator`
 //! to fill in this file with optimized constraints.
 
+use crate::air::memory_layout::DynamicTasmConstraintEvaluationMemoryLayout;
+use crate::air::memory_layout::StaticTasmConstraintEvaluationMemoryLayout;
 use crate::instruction::LabelledInstruction;
 use crate::table::constraints::ERROR_MESSAGE_GENERATE_CONSTRAINTS;
-use crate::table::StaticTasmConstraintEvaluationMemoryLayout;
 
 pub fn static_air_constraint_evaluation_tasm(
     _: StaticTasmConstraintEvaluationMemoryLayout,
@@ -13,7 +14,7 @@ pub fn static_air_constraint_evaluation_tasm(
 }
 
 pub fn dynamic_air_constraint_evaluation_tasm(
-    _: StaticTasmConstraintEvaluationMemoryLayout,
+    _: DynamicTasmConstraintEvaluationMemoryLayout,
 ) -> Vec<LabelledInstruction> {
     panic!("{ERROR_MESSAGE_GENERATE_CONSTRAINTS}");
 }
